flask_api/requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
easyocr==1.7.2
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
h11==0.16.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.32.3
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pyclipper==1.3.0.post6
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
python-bidi==0.6.6
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
selenium==4.34.0
shapely==2.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tifffile==2025.5.10
torch==2.2.2
torchvision==0.17.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
